feat: add useIsMounted
