+++ examples/extension-dapp/src/hooks/useProvider.ts
@@ -2,29 +2,23 @@ import {
   type LightClientProvider,
   getLightClientProvider,
 } from "@polkadot-api/light-client-extension-helpers/web-page"
-import { useEffect, useRef, useState } from "react"
+import { useEffect, useState } from "react"
+import { useIsMounted } from "./useIsMounted"
 
 const providers = new Map<string, Promise<LightClientProvider>>()
 
 export const useLightClientProvider = (channelId: string) => {
   const [provider, setProvider] = useState<LightClientProvider>()
-  const isMounted = useRef(false)
-
-  useEffect(() => {
-    isMounted.current = true
-    return () => {
-      isMounted.current = false
-    }
-  }, [])
+  const isMounted = useIsMounted()
 
   useEffect(() => {
     if (!providers.has(channelId))
       providers.set(channelId, getLightClientProvider(channelId))
     providers.get(channelId)?.then((provider) => {
-      if (!isMounted.current) return
+      if (!isMounted()) return
       setProvider(provider)
     })
-  }, [channelId])
+  }, [channelId, isMounted])
 
   return { provider }
 }
